Refactored navigation
